requirements.txt
elasticsearch==8.15.1
requests==2.32.3
openai==1.52.2
anthropic==0.37.1
google-cloud==0.34.0
google-auth==2.35.0
llama-index-core==0.11.20
